refactor(filter): merged MultipleSelect/SingleSelect into 1 class - renamed selectFilter to nativeSelectFilter so that I can use selectFilter name for something else (merged class)
