+++ aurelia-slickgrid/src/aurelia-slickgrid/filters/singleSelectFilter.ts
@@ -1,227 +1,14 @@
 import { I18N } from 'aurelia-i18n';
-import { inject } from 'aurelia-framework';
-import {
-  Column,
-  Filter,
-  FilterArguments,
-  FilterCallback,
-  GridOption,
-  MultipleSelectOption,
-  OperatorString,
-  OperatorType,
-  SearchTerm,
-  SelectOption
-} from './../models/index';
+import { inject, BindingEngine } from 'aurelia-framework';
+import { SelectFilter } from './selectFilter';
 import { CollectionService } from '../services/collection.service';
-import { htmlEncode } from '../services/utilities';
-import * as sanitizeHtml from 'sanitize-html';
-import * as $ from 'jquery';
 
-@inject(CollectionService, I18N)
-export class SingleSelectFilter implements Filter {
-  $filterElm: any;
-  grid: any;
-  searchTerms: SearchTerm[];
-  columnDef: Column;
-  callback: FilterCallback;
-  defaultOptions: MultipleSelectOption;
-  isFilled = false;
-  labelName: string;
-  labelPrefixName: string;
-  labelSuffixName: string;
-  valueName: string;
-  enableTranslateLabel = false;
-
-  constructor(private collectionService: CollectionService, private i18n: I18N) {
-    // default options used by this Filter, user can overwrite any of these by passing "otions"
-    this.defaultOptions = {
-      container: 'body',
-      filter: false,  // input search term on top of the select option list
-      maxHeight: 200,
-      single: true,
-      textTemplate: ($elm) => {
-        // render HTML code or not, by default it is sanitized and won't be rendered
-        const isRenderHtmlEnabled = this.columnDef && this.columnDef.filter && this.columnDef.filter.enableRenderHtml || false;
-        return isRenderHtmlEnabled ? $elm.text() : $elm.html();
-      },
-      onClose: () => {
-        const selectedItems = this.$filterElm.multipleSelect('getSelects');
-        let selectedItem = '';
-
-        if (Array.isArray(selectedItems) && selectedItems.length > 0) {
-          selectedItem = selectedItems[0] || null;
-          this.isFilled = true;
-          this.$filterElm.addClass('filled').siblings('div .search-filter').addClass('filled');
-        } else {
-          this.isFilled = false;
-          this.$filterElm.removeClass('filled').siblings('div .search-filter').removeClass('filled');
-        }
-
-        this.callback(undefined, { columnDef: this.columnDef, operator: this.operator, searchTerms: (selectedItem ? [selectedItem] : null) });
-      }
-    };
-  }
-
-  get operator(): OperatorType | OperatorString {
-    return (this.columnDef && this.columnDef.filter && this.columnDef.filter.operator) || OperatorType.equal;
-  }
-
-  /** Getter for the Grid Options pulled through the Grid Object */
-  private get gridOptions(): GridOption {
-    return (this.grid && this.grid.getOptions) ? this.grid.getOptions() : {};
-  }
-
+@inject(BindingEngine, CollectionService, I18N)
+export class SingleSelectFilter extends SelectFilter {
   /**
    * Initialize the Filter
    */
-  init(args: FilterArguments) {
-    if (!args) {
-      throw new Error('[Aurelia-SlickGrid] A filter must always have an "init()" with valid arguments.');
-    }
-    this.grid = args.grid;
-    this.callback = args.callback;
-    this.columnDef = args.columnDef;
-    this.searchTerms = args.searchTerms || [];
-
-    if (!this.grid || !this.columnDef || !this.columnDef.filter || !this.columnDef.filter.collection) {
-      throw new Error(`[Aurelia-SlickGrid] You need to pass a "collection" for the MultipleSelect Filter to work correctly. Also each option should include a value/label pair (or value/labelKey when using Locale). For example: { filter: { model: Filters.singleSelect, collection: [{ value: true, label: 'True' }, { value: false, label: 'False'}] } }`);
-    }
-
-    this.enableTranslateLabel = this.columnDef.filter.enableTranslateLabel || false;
-    this.labelName = (this.columnDef.filter.customStructure) ? this.columnDef.filter.customStructure.label : 'label';
-    this.labelPrefixName = (this.columnDef.filter.customStructure) ? this.columnDef.filter.customStructure.labelPrefix : 'labelPrefix';
-    this.labelSuffixName = (this.columnDef.filter.customStructure) ? this.columnDef.filter.customStructure.labelSuffix : 'labelSuffix';
-    this.valueName = (this.columnDef.filter.customStructure) ? this.columnDef.filter.customStructure.value : 'value';
-
-    let newCollection = this.columnDef.filter.collection || [];
-
-    // user might want to filter certain items of the collection
-    if (this.gridOptions.params && this.columnDef.filter.collectionFilterBy) {
-      const filterBy = this.columnDef.filter.collectionFilterBy;
-      newCollection = this.collectionService.filterCollection(newCollection, filterBy);
-    }
-
-    // user might want to sort the collection
-    if (this.columnDef.filter && this.columnDef.filter.collectionSortBy) {
-      const sortBy = this.columnDef.filter.collectionSortBy;
-      newCollection = this.collectionService.sortCollection(newCollection, sortBy, this.enableTranslateLabel);
-    }
-
-    let searchTerm = (Array.isArray(this.searchTerms) && this.searchTerms[0]) || '';
-    if (typeof searchTerm === 'boolean' || typeof searchTerm === 'number') {
-      searchTerm = `${searchTerm}`;
-    }
-
-    // step 1, create HTML string template
-    const filterTemplate = this.buildTemplateHtmlString(newCollection || [], searchTerm);
-
-    // step 2, create the DOM Element of the filter & pre-load search term
-    this.createDomElement(filterTemplate);
-  }
-
-  /**
-   * Clear the filter values
-   */
-  clear() {
-    if (this.$filterElm && this.$filterElm.multipleSelect) {
-      // reload the filter element by it's id, to make sure it's still a valid element (because of some issue in the GraphQL example)
-      this.$filterElm.multipleSelect('setSelects', []);
-      this.$filterElm.removeClass('filled');
-      this.searchTerms = [];
-      this.callback(undefined, { columnDef: this.columnDef, clearFilterTriggered: true });
-    }
-  }
-
-  /**
-   * destroy the filter
-   */
-  destroy() {
-    if (this.$filterElm) {
-      this.$filterElm.off().remove();
-    }
-  }
-
-  /**
-   * Set value(s) on the DOM element
-   */
-  setValues(values: SearchTerm[]) {
-    if (values) {
-      values = Array.isArray(values) ? values : [values];
-      this.$filterElm.multipleSelect('setSelects', values);
-    }
-  }
-
-  //
-  // private functions
-  // ------------------
-
-  /**
-   * Create the HTML template as a string
-   */
-  private buildTemplateHtmlString(optionCollection: any[], searchTerm?: SearchTerm) {
-    let options = '';
-    const isAddingSpaceBetweenLabels = this.columnDef && this.columnDef.filter && this.columnDef.filter.customStructure && this.columnDef.filter.customStructure.addSpaceBetweenLabels || false;
-    const isRenderHtmlEnabled = this.columnDef && this.columnDef.filter && this.columnDef.filter.enableRenderHtml || false;
-    const sanitizedOptions = this.gridOptions && this.gridOptions.sanitizeHtmlOptions || {};
-
-    optionCollection.forEach((option: SelectOption) => {
-      if (!option || (option[this.labelName] === undefined && option.labelKey === undefined)) {
-        throw new Error(`A collection with value/label (or value/labelKey when using Locale) is required to populate the Select list, for example: { filter: { model: Filter.singleSelect, collection: [ { value: '1', label: 'One' } ] } }`);
-      }
-
-      const labelKey = (option.labelKey || option[this.labelName]) as string;
-      const selected = (option[this.valueName] === searchTerm) ? 'selected' : '';
-      const labelText = ((option.labelKey || this.enableTranslateLabel) && this.i18n && typeof this.i18n.tr === 'function') ? this.i18n.tr(labelKey || ' ') : labelKey;
-      const prefixText = option[this.labelPrefixName] || '';
-      const suffixText = option[this.labelSuffixName] || '';
-      let optionText = isAddingSpaceBetweenLabels ? `${prefixText} ${labelText} ${suffixText}` : (prefixText + labelText + suffixText);
-
-      // if user specifically wants to render html text, he needs to opt-in else it will stripped out by default
-      // also, the 3rd party lib will saninitze any html code unless it's encoded, so we'll do that
-      if (isRenderHtmlEnabled) {
-        // sanitize any unauthorized html tags like script and others
-        // for the remaining allowed tags we'll permit all attributes
-        const sanitizeText = sanitizeHtml(optionText, sanitizedOptions);
-        optionText = htmlEncode(sanitizeText);
-      }
-
-      // html text of each select option
-      options += `<option value="${option[this.valueName]}" ${selected}>${optionText}</option>`;
-
-      // if there's a search term, we will add the "filled" class for styling purposes
-      if (selected) {
-        this.isFilled = true;
-      }
-    });
-
-    return `<select class="ms-filter search-filter">${options}</select>`;
-  }
-
-  /**
-   * From the html template string, create a DOM element
-   * Subscribe to the onClose event and run the callback when that happens
-   * @param filterTemplate
-   */
-  private createDomElement(filterTemplate: string) {
-    const $headerElm = this.grid.getHeaderRowColumn(this.columnDef.id);
-    $($headerElm).empty();
-
-    // create the DOM element & add an ID and filter class
-    this.$filterElm = $(filterTemplate);
-    if (typeof this.$filterElm.multipleSelect !== 'function') {
-      throw new Error(`multiple-select.js was not found, make sure to read the HOWTO Wiki on how to install it`);
-    }
-    this.$filterElm.attr('id', `filter-${this.columnDef.id}`);
-    this.$filterElm.data('columnId', this.columnDef.id);
-
-    // append the new DOM element to the header row
-    if (this.$filterElm && typeof this.$filterElm.appendTo === 'function') {
-      this.$filterElm.appendTo($headerElm);
-    }
-
-    // merge options & attach multiSelect
-    const filterOptions = (this.columnDef.filter) ? this.columnDef.filter.filterOptions : {};
-    const options: MultipleSelectOption = { ...this.defaultOptions, ...filterOptions };
-    this.$filterElm = this.$filterElm.multipleSelect(options);
+  constructor(protected bindingEngine: BindingEngine, protected collectionService: CollectionService, protected i18n: I18N) {
+    super(bindingEngine, collectionService, i18n, false);
   }
 }
